refactor: Move dev mode handler initialization to DevModeHandlerManager
This is needed so Spring can call it from its own init mechanism
